split 1 type per file and rename folder/csproj/namespace to FreshTinyIoC because it's impossible to understand anything in that project, also remove useless #undef expect in the only class that uses it
