+++ src/FreshTinyIoC/FreshTinyIoC.csproj
@@ -8,14 +8,4 @@
     <Deterministic>false</Deterministic>
     <ReleaseVersion>2.2.0</ReleaseVersion>
   </PropertyGroup>
-  <ItemGroup>
-    <Compile Remove="backup-2019-01-17-093141\**" />
-    <EmbeddedResource Remove="backup-2019-01-17-093141\**" />
-    <None Remove="backup-2019-01-17-093141\**" />
-  </ItemGroup>
-  <ItemGroup>
-  </ItemGroup>
-  
-  <ItemGroup>
-  </ItemGroup>
 </Project>
