+++ src/FreshTinyIoC/IRegisterOptions.cs
